refactor: separate withdraw confirm component
